Split Lightning dashboard tiles
