+++ BTCPayServer/Components/StoreLightningBalance/Default.cshtml
@@ -1,6 +1,3 @@
-@using BTCPayServer.TagHelpers
-@using Microsoft.AspNetCore.Mvc.TagHelpers
-@using BTCPayServer.Client
 @using BTCPayServer.Lightning
 @model BTCPayServer.Components.StoreLightningBalance.StoreLightningBalanceViewModel
 
@@ -16,7 +13,7 @@
     }
     else
     {
-        <div class="balances d-flex flex-wrap gap-3">
+        <div class="balances d-flex flex-wrap gap-4">
             <div class="balance me-3">
                 <h3 class="d-inline-block me-1">@Model.TotalOffchain</h3>
                 <span class="text-secondary fw-semibold">@Model.CryptoCode in channels</span>
@@ -78,46 +75,6 @@
                     </div>
                 }
             </div>
-        </div>
-    }
-    
-    @if (Model.Services != null && Model.Services.Any())
-    {
-        <header class="mt-5 mb-4">
-            <h6>Lightning Services</h6>
-            <a
-                asp-controller="UIPublicLightningNodeInfo"
-                asp-action="ShowLightningNodeInfo"app-top-items
-                asp-route-cryptoCode="@Model.CryptoCode"
-                asp-route-storeId="@Model.Store.Id"
-                target="_blank"
-                id="PublicNodeInfo">
-                Node Info
-            </a>
-        </header>
-        <div id="Services" class="services-list">
-            @foreach (var service in Model.Services)
-            {
-                @if (!string.IsNullOrEmpty(service.Error))
-                {
-                    <div class="service" id="@($"Service-{service.ServiceName}")">
-                        <img src="@($"~/img/{service.Type.ToLower()}.png")" asp-append-version="true" alt="@service.DisplayName" />
-                        <small class="d-block mt-3 lh-sm fw-semibold text-danger">@service.Error</small>
-                    </div>
-                }
-                else if (string.IsNullOrEmpty(service.Link))
-                {
-                    <a asp-controller="UIServer" asp-action="Service" asp-route-serviceName="@service.ServiceName" asp-route-cryptoCode="@service.CryptoCode" class="service" id="@($"Service-{service.ServiceName}")">
-                        <img src="@($"~/img/{service.Type.ToLower()}.png")" asp-append-version="true" alt="@service.DisplayName" />
-                    </a>
-                }
-                else
-                {
-                    <a href="@service.Link" target="_blank" rel="noreferrer noopener" class="service" id="@($"Service-{service.ServiceName}")">
-                        <img src="@($"~/img/{service.Type.ToLower()}.png")" asp-append-version="true" alt="@service.DisplayName" />
-                    </a>
-                }
-            }
         </div>
     }
 </div>
