+++ BTCPayServer/Views/UIStores/Dashboard.cshtml
@@ -88,6 +88,7 @@
         @if (Model.LightningEnabled)
         {
             <vc:store-lightning-balance store="@store"/>
+            <vc:store-lightning-services store="@store"/>
         }
         @if (Model.WalletEnabled)
         {
